Handle setup by state; make more user-friendly
Set up the database by state, with command-line options necessary
to choose which state (or both) the scripts will work on. Additionally,
reset the database by state, rather than completely dropping and
recreating it. Finally, usage of the setup script is now documented,
and if the script is run incorrectly it will exit after showing the
usage text.